SPARK-16972: Move DriverEndpoint out of CoarseGrainedSchedulerBackend
