+++ yarn/src/main/scala/org/apache/spark/scheduler/cluster/YarnSchedulerBackend.scala
@@ -163,7 +163,7 @@ private[spark] abstract class YarnSchedulerBackend(
   }
 
   override def createDriverEndpoint(properties: Seq[(String, String)]): DriverEndpoint = {
-    new YarnDriverEndpoint(rpcEnv, properties)
+    new YarnDriverEndpoint(this, scheduler, rpcEnv, properties)
   }
 
   /**
@@ -181,8 +181,10 @@ private[spark] abstract class YarnSchedulerBackend(
    * This endpoint communicates with the executors and queries the AM for an executor's exit
    * status when the executor is disconnected.
    */
-  private class YarnDriverEndpoint(rpcEnv: RpcEnv, sparkProperties: Seq[(String, String)])
-      extends DriverEndpoint(rpcEnv, sparkProperties) {
+  private class YarnDriverEndpoint(backend: CoarseGrainedSchedulerBackend,
+                                   scheduler: TaskSchedulerImpl,
+                                   rpcEnv: RpcEnv, sparkProperties: Seq[(String, String)])
+      extends DriverEndpoint(backend, scheduler, rpcEnv, sparkProperties) {
 
     /**
      * When onDisconnected is received at the driver endpoint, the superclass DriverEndpoint
